# judesys/ClientApp/swagger.json
{"swagger":"2.0","info":{"description":"Judesys API endpoints testing","version":"0.1","title":"Judesys API","termsOfService":"http://abobus.com","contact":{"name":"vlaryz@ktu.lt"},"license":{"name":"MIT License","url":"vlaryz@ktu.lt"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"ticket-controller","description":"Ticket Controller"},{"name":"auth-controller","description":"Auth Controller"},{"name":"event-controller","description":"Event Controller"},{"name":"city-controller","description":"City Controller"}],"paths":{"/api/cities/":{"get":{"tags":["city-controller"],"summary":"getAllCities","operationId":"getAllCitiesUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["city-controller"],"summary":"addCity","operationId":"addCityUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"city","description":"city","required":true,"schema":{"$ref":"#/definitions/CityRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CityResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/cities/{cityId}/events/":{"get":{"tags":["event-controller"],"summary":"getAllEvents","operationId":"getAllEventsUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"cityId","in":"path","description":"cityId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/EventResponse"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["event-controller"],"summary":"addEvent","operationId":"addEventUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"cityId","in":"path","description":"cityId","required":true,"type":"integer","format":"int64"},{"in":"body","name":"event","description":"event","required":true,"schema":{"$ref":"#/definitions/EventRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/EventResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/cities/{cityId}/events/{eventId}/tickets/":{"get":{"tags":["ticket-controller"],"summary":"getAllTickets","operationId":"getAllTicketsUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"eventId","in":"path","description":"eventId","required":true,"type":"integer","format":"int64"},{"name":"cityId","in":"path","description":"cityId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/TicketResponse"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["ticket-controller"],"summary":"addTicket","operationId":"addTicketUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"eventId","in":"path","description":"eventId","required":true,"type":"integer","format":"int64"},{"name":"cityId","in":"path","description":"cityId","required":true,"type":"integer","format":"int64"},{"in":"body","name":"ticketRequest","description":"ticketRequest","required":true,"schema":{"$ref":"#/definitions/TicketRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TicketResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/cities/{cityId}/events/{eventId}/tickets/byUser":{"get":{"tags":["ticket-controller"],"summary":"getAllTicketsByUser","operationId":"getAllTicketsByUserUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"eventId","in":"path","description":"eventId","required":true,"type":"integer","format":"int64"},{"name":"cityId","in":"path","description":"cityId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/TicketDetailsResponse"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/cities/{cityId}/events/{eventId}/tickets/{id}":{"get":{"tags":["ticket-controller"],"summary":"getTicket","operationId":"getTicketUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"name":"eventId","in":"path","description":"eventId","required":true,"type":"integer","format":"int64"},{"name":"cityId","in":"path","description":"cityId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TicketResponse"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["ticket-controller"],"summary":"updateTicket","operationId":"updateTicketUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"name":"eventId","in":"path","description":"eventId","required":true,"type":"integer","format":"int64"},{"name":"cityId","in":"path","description":"cityId","required":true,"type":"integer","format":"int64"},{"in":"body","name":"ticketRequest","description":"ticketRequest","required":true,"schema":{"$ref":"#/definitions/TicketRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TicketResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["ticket-controller"],"summary":"deleteTicket","operationId":"deleteTicketUsingDELETE","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"name":"eventId","in":"path","description":"eventId","required":true,"type":"integer","format":"int64"},{"name":"cityId","in":"path","description":"cityId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}}},"/api/cities/{cityId}/events/{id}":{"get":{"tags":["event-controller"],"summary":"getEvent","operationId":"getEventUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"cityId","in":"path","description":"cityId","required":true,"type":"integer","format":"int64"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/EventResponse"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["event-controller"],"summary":"updateEvent","operationId":"updateEventUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"cityId","in":"path","description":"cityId","required":true,"type":"integer","format":"int64"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"in":"body","name":"event","description":"event","required":true,"schema":{"$ref":"#/definitions/EventRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/EventResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["event-controller"],"summary":"deleteEvent","operationId":"deleteEventUsingDELETE","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"cityId","in":"path","description":"cityId","required":true,"type":"integer","format":"int64"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}}},"/api/cities/{id}":{"get":{"tags":["city-controller"],"summary":"getCity","operationId":"getCityUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CityResponse"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["city-controller"],"summary":"updateCity","operationId":"updateCityUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"in":"body","name":"city","description":"city","required":true,"schema":{"$ref":"#/definitions/CityRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CityResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["city-controller"],"summary":"deleteCity","operationId":"deleteCityUsingDELETE","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}}},"/api/users/":{"get":{"tags":["auth-controller"],"summary":"getUsers","operationId":"getUsersUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["auth-controller"],"summary":"registerUser","operationId":"registerUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"registerUserRequest","description":"registerUserRequest","required":true,"schema":{"$ref":"#/definitions/RegisterUserRequest"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/users/auth":{"post":{"tags":["auth-controller"],"summary":"getAuthenticationToken","operationId":"getAuthenticationTokenUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"yourRequestDTO","description":"yourRequestDTO","required":true,"schema":{"$ref":"#/definitions/LoginRequest"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/users/roles/":{"post":{"tags":["auth-controller"],"summary":"saveRole","operationId":"saveRoleUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"roleRequest","description":"roleRequest","required":true,"schema":{"$ref":"#/definitions/RoleRequest"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["auth-controller"],"summary":"addRoleToUser","operationId":"addRoleToUserUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userRoleRequest","description":"userRoleRequest","required":true,"schema":{"$ref":"#/definitions/UserRoleRequest"}}],"responses":{"201":{"description":"Created"},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/users/token/refresh":{"get":{"tags":["auth-controller"],"summary":"refreshToken","operationId":"refreshTokenUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK"},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"Ticket":{"type":"object","properties":{"boughtBy":{"type":"integer","format":"int64"},"createdBy":{"type":"integer","format":"int64"},"id":{"type":"integer","format":"int64"},"price":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["BASIC","VIP"]}}},"User":{"type":"object","properties":{"email":{"type":"string"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"password":{"type":"string"},"roles":{"$ref":"#/definitions/Collection«Role»"},"surname":{"type":"string"},"username":{"type":"string"}}},"RegisterUserRequest":{"type":"object","properties":{"email":{"type":"string"},"name":{"type":"string"},"password":{"type":"string"},"surname":{"type":"string"},"user":{"$ref":"#/definitions/User"},"userName":{"type":"string"}}},"Collection«Role»":{"type":"object"},"EventRequest":{"type":"object","properties":{"event":{"$ref":"#/definitions/Event"},"free":{"type":"boolean"},"longDescription":{"type":"string"},"name":{"type":"string"},"shortDescription":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"password":{"type":"string"},"username":{"type":"string"}}},"City":{"type":"object","properties":{"createdBy":{"type":"integer","format":"int64"},"events":{"type":"array","items":{"$ref":"#/definitions/Event"}},"id":{"type":"integer","format":"int64"},"image":{"type":"string"},"name":{"type":"string"}}},"TicketDetailsResponse":{"type":"object","properties":{"boughtBy":{"type":"integer","format":"int64"},"city":{"type":"string"},"event":{"type":"string"},"id":{"type":"integer","format":"int64"},"price":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["BASIC","VIP"]}}},"TicketRequest":{"type":"object","properties":{"price":{"type":"integer","format":"int64"},"ticket":{"$ref":"#/definitions/Ticket"},"type":{"type":"string","enum":["BASIC","VIP"]}}},"RoleRequest":{"type":"object","properties":{"role":{"$ref":"#/definitions/Role"},"roleName":{"type":"string"}}},"Role":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"CityResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"image":{"type":"string"},"name":{"type":"string"}}},"UserRoleRequest":{"type":"object","properties":{"role":{"type":"string"},"username":{"type":"string"}}},"EventResponse":{"type":"object","properties":{"free":{"type":"boolean"},"id":{"type":"integer","format":"int64"},"longDescr":{"type":"string"},"name":{"type":"string"},"shortDescr":{"type":"string"}}},"CityRequest":{"type":"object","properties":{"city":{"$ref":"#/definitions/City"},"image":{"type":"string"},"name":{"type":"string"}}},"TicketResponse":{"type":"object","properties":{"boughtBy":{"type":"integer","format":"int64"},"id":{"type":"integer","format":"int64"},"price":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["BASIC","VIP"]}}},"Event":{"type":"object","properties":{"createdBy":{"type":"integer","format":"int64"},"free":{"type":"boolean"},"id":{"type":"integer","format":"int64"},"longDesription":{"type":"string"},"name":{"type":"string"},"shortDesription":{"type":"string"},"tickets":{"type":"array","items":{"$ref":"#/definitions/Ticket"}}}}}}